# backend/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stellar Payment API",
    "version": "0.1.0",
    "description": "API for creating and verifying Stellar network payments"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "paths": {
...
    "/api/create-payment": {
      "post": {
        "summary": "Create a new payment session request",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "schema": {
              "type": "string"
            },
            "description": "Optional unique key for idempotent requests. Use UUID or request ID. Responses are cached for 24 hours."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount",
                  "asset",
                  "recipient"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "Payment amount (must be positive and at least 0.01 XLM for native payments)"
                  },
                  "asset": {
                    "type": "string",
                    "description": "Asset code (e.g. XLM, USDC)"
                  },
                  "asset_issuer": {
                    "type": "string",
                    "description": "Asset issuer (required for non-native assets)"
                  },
                  "recipient": {
                    "type": "string",
                    "description": "Stellar address of the recipient"
                  },
                  "merchant_id": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "memo": {
                    "type": "string"
                  },
                  "memo_type": {
                    "type": "string",
                    "enum": [
                      "text",
                      "id",
                      "hash",
                      "return"
                    ]
                  },
                  "webhook_url": {
                    "type": "string"
                  },
                  "branding_overrides": {
                    "type": "object",
                    "properties": {
                      "primary_color": {
                        "type": "string",
                        "example": "#5ef2c0"
                      },
                      "secondary_color": {
                        "type": "string",
                        "example": "#b8ffe2"
                      },
                      "background_color": {
                        "type": "string",
                        "example": "#050608"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Duplicate request — cached response returned from idempotency key",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payment_id": {
                      "type": "string"
                    },
                    "payment_link": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Payment created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payment_id": {
                      "type": "string"
                    },
                    "payment_link": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "branding_config": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid Idempotency-Key"
          },
          "429": {
            "description": "Too many requests"
          }
        }
      }
    },
    "/api/payment-status/{id}": {
      "get": {
        "summary": "Get the status of a payment",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Payment ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payment": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Payment not found"
          }
        }
      }
    },
    "/api/verify-payment/{id}": {
      "post": {
        "summary": "Verify a payment on the Stellar network",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Payment ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "pending",
                        "confirmed"
                      ]
                    },
                    "tx_id": {
                      "type": "string"
                    },
                    "webhook": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Payment not found"
          }
        }
      }
    },
    "/api/payments": {
      "get": {
        "summary": "Get paginated list of payments for the authenticated merchant",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number (1-indexed)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of results per page (max 100)"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated payments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payments": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "total_count": {
                      "type": "integer"
                    },
                    "total_pages": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        }
      }
    },
...
    "/api/payments/{id}/refund": {
      "post": {
        "summary": "Generate a refund transaction for a confirmed payment",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Payment ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Refund transaction XDR",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "xdr": {
                      "type": "string",
                      "description": "Transaction XDR to sign and submit"
                    },
                    "hash": {
                      "type": "string",
                      "description": "Transaction hash"
                    },
                    "instructions": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Payment not eligible for refund"
          },
          "404": {
            "description": "Payment not found"
          }
        }
      }
    },
    "/api/payments/{id}/refund/confirm": {
      "post": {
        "summary": "Confirm a refund transaction has been submitted",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Payment ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tx_hash"
                ],
                "properties": {
                  "tx_hash": {
                    "type": "string",
                    "description": "Submitted refund transaction hash"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refund confirmed"
          },
          "404": {
            "description": "Payment not found"
          }
        }
      }
    },
    "/api/path-payment-quote/{id}": {
      "get": {
        "summary": "Get a path payment quote for a payment session",
        "description": "Returns the estimated send amount if the customer wants to pay with a different asset than the merchant expects.",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Payment ID"
          },
          {
            "in": "query",
            "name": "source_asset",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Asset code the customer wants to send (e.g. XLM)"
          },
          {
            "in": "query",
            "name": "source_asset_issuer",
            "schema": {
              "type": "string"
            },
            "description": "Issuer of the source asset (required if not XLM)"
          },
          {
            "in": "query",
            "name": "source_account",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer's Stellar public key"
          }
        ],
        "responses": {
          "200": {
            "description": "Path payment quote"
          },
          "400": {
            "description": "Missing parameters or same asset"
          },
          "404": {
            "description": "Payment not found or no path available"
          }
        }
      }
    },
    "/api/webhooks/test": {
      "post": {
        "summary": "Send a test webhook to the merchant's stored webhook URL",
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Test webhook dispatched",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "body": {
                      "type": "string"
                    },
                    "signed": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No webhook URL configured"
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        }
      }
    }
  },
...
}